pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "truth-matrix"
version = "0.1.0"
description = "Truth Matrix escape-room puzzle (Raspberry Pi + logic solver)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "gpiozero>=2.0",
  "jsonschema>=4.0",
  "requests>=2.28",
]

[project.optional-dependencies]
display = ["luma.oled>=3.12", "pillow>=9.0"]

[tool.setuptools.packages.find]
where = ["src"]
